LayoutTests/editing/pasteboard/copy-paste-first-line-in-textarea-expected.txt


PASS selectFirstLineIncludingLineBreak(); copyPaste(); textarea.selectionEnd is offsetAfterLineBreak
